refactor(bpmn): follow the flow suite's shape — no judge, shared helpers
Four conformance changes against the #2601 standard.

1. Removed `agent_judge` and the `checker_context` block it forced. The
   flow-side twin of this exact scenario has no judge: its persona withholds
   policy facts and the hidden scenarios fail if the agent never elicited
   them, so the interview is graded by its consequences. Ours was the repo's
   only agent_judge, and it was the sole reason for the route pin that
   blocked every run. Weight 25.5 -> 21.5; live-derived weight unchanged at
   12.0, now 56% of the total.

2. Extracted `_shared/bpmn_live.py` (379 lines, 20 helpers) to play the role
   `flow_check.py` plays for the flow suite, and pointed both this task's
   checkers at it. The package checker's `parse_json_output` was a third
   verbatim copy; it now delegates while keeping its own FAIL: convention.

3. Ported `check_bindings_no_stubs`'s non-stub-UUID rule. The pack criterion
   accepted any truthy Connection key, so an all-zero UUID -- exactly what
   the CLI writes for an unresolved binding -- passed a criterion meant to
   prove real wiring.

4. Caught during the extraction: `run_cli` moved to the shared module, so it
   reads the shared `ACTIVE_CLI_DEADLINE`. Assigning the task-local copy
   would have silently disabled deadline capping again -- the same bug fixed
   one commit ago, reintroduced by the move. Both assignment sites (main and
   the post_run sweep) now target `bpmn_live.ACTIVE_CLI_DEADLINE`.

Deliberately NOT done: `debug/live_debug_e2e/check_live_debug.py` keeps its
private helpers. They look like duplicates but have different contracts --
`_parse_json_tolerant` returns None and skips a CLI banner where the shared
one raises, `_find_ci_key` is a recursive generator where `get_ci` is a
single-level lookup, and `_run_uip` retries twice where `run_cli` does not.
Converting it would change behaviour in a separate passing path-to-ga task I
cannot run, so it belongs in its own PR.

  behavior checker  2404 -> 2097
  task yaml          439 ->  365
  _shared/bpmn_live.py    new, 379 (amortizable across future BPMN live evals)
